+++ internal/oauth2/controller/oauth2_controller_test.go
@@ -20,7 +20,7 @@ import (
 	oauth2Service "github.com/rushairer/gosso/internal/oauth2/service"
 	tokenDomain "github.com/rushairer/gosso/internal/token/domain"
 
-	"github.com/rushairer/gosso/internal/auth/middleware"
+	"github.com/rushairer/gosso/middleware"
 )
 
 // ──────────────────────────────────────────────